Own custom type errors at resolution boundaries
- Separate template, constructor, and field validation failures.
- Preserve structured custom identities and mismatch context.
- Propagate failures from the lookup that owns each boundary.
