+++ src/planner/context.rs
@@ -1,30 +1,28 @@
+mod custom;
+
 use crate::plan::{
     BitArrayFunctionLocalId, BitArrayListLocalId, BitArrayLocalId, BoolFunctionLocalId,
-    BoolListLocalId, BoolLocalId, CaptureArg, CustomConstructor, CustomConstructorField,
-    CustomExpr, CustomFieldAccess, CustomFunctionLocal, CustomFunctionLocalId, CustomFunctionType,
-    CustomListLocalId, CustomLocalId, CustomTypeDefinition, CustomTypeTemplate,
-    ExternalFunctionLocal, ExternalFunctionLocalId, ExternalFunctionType, ExternalListLocalId,
-    ExternalLocalId, FloatFunctionLocalId, FloatListLocalId, FloatLocalId, FunctionFunctionLocal,
+    BoolListLocalId, BoolLocalId, CaptureArg, CustomFunctionLocal, CustomFunctionLocalId,
+    CustomFunctionType, CustomListLocalId, CustomLocalId, CustomValueShape, ExternalFunctionLocal,
+    ExternalFunctionLocalId, ExternalFunctionType, ExternalListLocalId, ExternalLocalId,
+    FloatFunctionLocalId, FloatListLocalId, FloatLocalId, FunctionFunctionLocal,
     FunctionFunctionLocalId, FunctionFunctionType, FunctionListLocalId, FunctionReference,
-    FunctionTemplate, FunctionTemplateSignature, FunctionType, IntFunctionLocalId, IntListLocalId,
-    IntLocalId, ListExpr, ListFunctionLocal, ListListLocalId, ListLocal, ListLocalExpr, LocalId,
-    ModuleId, NilFunctionLocalId, NilListLocalId, NilLocalId, PanicSite, ParamBinding, ParamLocal,
-    ParamSlot, StringFunctionLocalId, StringListLocalId, StringLocalId, TupleFunctionLocalId,
-    TupleListLocalId, TupleLocalId, UtfCodepointFunctionLocalId, UtfCodepointListLocalId,
-    UtfCodepointLocalId, ValueType,
-};
-use crate::plan::{
-    CustomConstructorRefinement, CustomType, CustomValueShape, FunctionShape, ValueShape,
+    FunctionShape, FunctionTemplate, FunctionTemplateSignature, FunctionType, IntFunctionLocalId,
+    IntListLocalId, IntLocalId, ListExpr, ListFunctionLocal, ListListLocalId, ListLocal,
+    ListLocalExpr, LocalId, ModuleId, NilFunctionLocalId, NilListLocalId, NilLocalId, PanicSite,
+    ParamBinding, ParamLocal, ParamSlot, StringFunctionLocalId, StringListLocalId, StringLocalId,
+    TupleFunctionLocalId, TupleListLocalId, TupleLocalId, UtfCodepointFunctionLocalId,
+    UtfCodepointListLocalId, UtfCodepointLocalId, ValueShape, ValueType,
 };
 use crate::planner::error::{
-    InvalidCustomTypeReason, InvalidExpressionShapeKind, InvalidModuleReferenceReason,
-    InvalidTypedAstReason, PlanError,
+    InvalidExpressionShapeKind, InvalidModuleReferenceReason, InvalidTypedAstReason, PlanError,
 };
 use ecow::EcoString;
-use gleam_core::type_::{
-    PRELUDE_MODULE_NAME, PatternConstructor, Type, ValueConstructor, ValueConstructorVariant,
-};
+use gleam_core::type_::Type;
 use std::collections::HashMap;
+
+#[cfg(test)]
+use crate::plan::CustomTypeDefinition;
 
 #[derive(Clone)]
 pub(super) struct FunctionInfo {
@@ -197,27 +195,6 @@ impl FunctionParam {
     }
 }
 
-#[derive(Debug, Clone, PartialEq)]
-pub(super) struct ResolvedCustomConstructor {
-    constructor: CustomConstructor,
-    constructor_count: usize,
-    source_shape: CustomValueShape,
-}
-
-impl ResolvedCustomConstructor {
-    pub(super) fn constructor_count(&self) -> usize {
-        self.constructor_count
-    }
-
-    pub(super) fn source_shape(&self) -> &CustomValueShape {
-        &self.source_shape
-    }
-
-    pub(super) fn into_constructor(self) -> CustomConstructor {
-        self.constructor
-    }
-}
-
 pub(super) struct PlanContext<'a> {
     pub(super) module_name: &'a EcoString,
     current_function: EcoString,
@@ -328,256 +305,6 @@ fn invalid_local_shape() -> PlanError {
             kind: InvalidExpressionShapeKind::Invalid,
         },
     }
-}
-
-fn instantiate_custom_type_template(
-    template: &CustomTypeTemplate,
-    custom_type: &CustomType,
-) -> Result<ValueType, PlanError> {
-    let type_ = match template {
-        CustomTypeTemplate::Int => ValueType::Int,
-        CustomTypeTemplate::Float => ValueType::Float,
-        CustomTypeTemplate::String => ValueType::String,
-        CustomTypeTemplate::BitArray => ValueType::BitArray,
-        CustomTypeTemplate::UtfCodepoint => ValueType::UtfCodepoint,
-        CustomTypeTemplate::Bool => ValueType::Bool,
-        CustomTypeTemplate::Nil => ValueType::Nil,
-        CustomTypeTemplate::Tuple(elements) => ValueType::Tuple(
-            elements
-                .iter()
-                .map(|element| instantiate_custom_type_template(element, custom_type))
-                .collect::<Result<Vec<_>, _>>()?,
-        ),
-        CustomTypeTemplate::List(element) => ValueType::List(Box::new(
-            instantiate_custom_type_template(element, custom_type)?,
-        )),
-        CustomTypeTemplate::Function { arguments, return_ } => {
-            ValueType::Function(Box::new(FunctionType::new(
-                arguments
-                    .iter()
-                    .map(|argument| instantiate_custom_type_template(argument, custom_type))
-                    .collect::<Result<Vec<_>, _>>()?,
-                instantiate_custom_type_template(return_, custom_type)?,
-            )))
-        }
-        CustomTypeTemplate::Custom { name, arguments } => ValueType::Custom(CustomType::new(
-            name.clone(),
-            arguments
-                .iter()
-                .map(|argument| instantiate_custom_type_template(argument, custom_type))
-                .collect::<Result<Vec<_>, _>>()?,
-        )),
-        CustomTypeTemplate::External { name, arguments } => {
-            ValueType::External(crate::plan::ExternalType::new(
-                name.clone(),
-                arguments
-                    .iter()
-                    .map(|argument| instantiate_custom_type_template(argument, custom_type))
-                    .collect::<Result<Vec<_>, _>>()?,
-            ))
-        }
-        CustomTypeTemplate::Parameter(parameter) => {
-            let Some(type_) = custom_type.arguments().get(parameter.0).cloned() else {
-                return Err(PlanError::InvalidTypedAst {
-                    reason: InvalidTypedAstReason::CustomType {
-                        name: custom_type.type_name().name().clone(),
-                        reason: InvalidCustomTypeReason::ParameterType,
-                    },
-                });
-            };
-            type_
-        }
-    };
-    Ok(type_)
-}
-
-fn instantiate_custom_shape_template(
-    template: &CustomTypeTemplate,
-    custom_shape: &CustomValueShape,
-) -> Result<ValueShape, PlanError> {
-    let shape = match template {
-        CustomTypeTemplate::Int => ValueShape::Int,
-        CustomTypeTemplate::Float => ValueShape::Float,
-        CustomTypeTemplate::String => ValueShape::String,
-        CustomTypeTemplate::BitArray => ValueShape::BitArray,
-        CustomTypeTemplate::UtfCodepoint => ValueShape::UtfCodepoint,
-        CustomTypeTemplate::Bool => ValueShape::Bool,
-        CustomTypeTemplate::Nil => ValueShape::Nil,
-        CustomTypeTemplate::Tuple(elements) => ValueShape::Tuple(
-            elements
-                .iter()
-                .map(|element| instantiate_custom_shape_template(element, custom_shape))
-                .collect::<Result<Vec<_>, _>>()?
-                .into_boxed_slice(),
-        ),
-        CustomTypeTemplate::List(item) => ValueShape::List(Box::new(
-            instantiate_custom_shape_template(item, custom_shape)?,
-        )),
-        CustomTypeTemplate::Function { arguments, return_ } => {
-            ValueShape::Function(Box::new(FunctionShape::new(
-                arguments
-                    .iter()
-                    .map(|argument| instantiate_custom_shape_template(argument, custom_shape))
-                    .collect::<Result<Vec<_>, _>>()?,
-                instantiate_custom_shape_template(return_, custom_shape)?,
-            )))
-        }
-        CustomTypeTemplate::Custom { name, arguments } => {
-            ValueShape::Custom(CustomValueShape::new(
-                name.clone(),
-                arguments
-                    .iter()
-                    .map(|argument| instantiate_custom_shape_template(argument, custom_shape))
-                    .collect::<Result<Vec<_>, _>>()?,
-                CustomConstructorRefinement::Any,
-            ))
-        }
-        CustomTypeTemplate::External { name, arguments } => {
-            ValueShape::External(crate::plan::ExternalValueShape::new(
-                name.clone(),
-                arguments
-                    .iter()
-                    .map(|argument| instantiate_custom_shape_template(argument, custom_shape))
-                    .collect::<Result<Vec<_>, _>>()?,
-            ))
-        }
-        CustomTypeTemplate::Parameter(parameter) => custom_shape
-            .arguments()
-            .get(parameter.0)
-            .cloned()
-            .ok_or_else(|| {
-                invalid_custom_type(custom_shape.type_(), InvalidCustomTypeReason::ParameterType)
-            })?,
-    };
-    Ok(shape)
-}
-
-fn invalid_custom_type(type_: &CustomType, reason: InvalidCustomTypeReason) -> PlanError {
-    PlanError::InvalidTypedAst {
-        reason: InvalidTypedAstReason::CustomType {
-            name: type_.type_name().name().clone(),
-            reason,
-        },
-    }
-}
-
-fn collect_parameter_shapes(
-    template: &CustomTypeTemplate,
-    shape: &ValueShape,
-    arguments: &mut [Option<ValueShape>],
-    owner: &CustomType,
-) -> Result<(), PlanError> {
-    match template {
-        CustomTypeTemplate::Parameter(parameter) => {
-            let Some(argument) = arguments.get_mut(parameter.0) else {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            };
-            *argument = Some(match argument.take() {
-                Some(previous) => previous.merge(shape).ok_or_else(|| {
-                    invalid_custom_type(owner, InvalidCustomTypeReason::ParameterType)
-                })?,
-                None => shape.clone(),
-            });
-        }
-        CustomTypeTemplate::Tuple(templates) => {
-            let ValueShape::Tuple(shapes) = shape else {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            };
-            if templates.len() != shapes.len() {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            }
-            for (template, shape) in templates.iter().zip(shapes.iter()) {
-                collect_parameter_shapes(template, shape, arguments, owner)?;
-            }
-        }
-        CustomTypeTemplate::List(template) => {
-            let ValueShape::List(shape) = shape else {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            };
-            collect_parameter_shapes(template, shape, arguments, owner)?;
-        }
-        CustomTypeTemplate::Function {
-            arguments: templates,
-            return_: return_template,
-        } => {
-            let ValueShape::Function(shape) = shape else {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            };
-            if templates.len() != shape.argument_shapes().len() {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            }
-            for (template, shape) in templates.iter().zip(shape.argument_shapes()) {
-                collect_parameter_shapes(template, shape, arguments, owner)?;
-            }
-            collect_parameter_shapes(return_template, shape.return_shape(), arguments, owner)?;
-        }
-        CustomTypeTemplate::Custom {
-            name,
-            arguments: templates,
-        } => {
-            let ValueShape::Custom(shape) = shape else {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            };
-            if shape.type_name() != name || templates.len() != shape.arguments().len() {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            }
-            for (template, shape) in templates.iter().zip(shape.arguments()) {
-                collect_parameter_shapes(template, shape, arguments, owner)?;
-            }
-        }
-        CustomTypeTemplate::External {
-            name,
-            arguments: templates,
-        } => {
-            let ValueShape::External(shape) = shape else {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            };
-            if shape.type_name() != name || templates.len() != shape.arguments().len() {
-                return Err(invalid_custom_type(
-                    owner,
-                    InvalidCustomTypeReason::ParameterType,
-                ));
-            }
-            for (template, shape) in templates.iter().zip(shape.arguments()) {
-                collect_parameter_shapes(template, shape, arguments, owner)?;
-            }
-        }
-        CustomTypeTemplate::Int
-        | CustomTypeTemplate::Float
-        | CustomTypeTemplate::String
-        | CustomTypeTemplate::BitArray
-        | CustomTypeTemplate::UtfCodepoint
-        | CustomTypeTemplate::Bool
-        | CustomTypeTemplate::Nil => {}
-    }
-    Ok(())
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -2341,408 +2068,6 @@ impl<'a> PlanContext<'a> {
         }
     }
 
-    fn validate_constructor_arity(
-        &self,
-        constructor: ResolvedCustomConstructor,
-        arity: usize,
-    ) -> Result<CustomConstructor, PlanError> {
-        let constructor = constructor.into_constructor();
-        if constructor.fields().len() != arity {
-            return Err(invalid_custom_type(
-                constructor.type_(),
-                InvalidCustomTypeReason::ConstructorArity,
-            ));
-        }
-        Ok(constructor)
-    }
-
-    fn custom_type_definition(
-        &self,
-        name: &crate::plan::CustomTypeName,
-    ) -> Option<&CustomTypeDefinition> {
-        match self.registry {
-            RegistryAccess::Program { registry } => registry.custom_type(name),
-            #[cfg(test)]
-            RegistryAccess::Local { custom_types, .. } => custom_types
-                .iter()
-                .find(|definition| definition.name() == name),
-        }
-    }
-
-    pub(super) fn custom_constructor(
-        &self,
-        constructor: &ValueConstructor,
-    ) -> Result<CustomConstructor, PlanError> {
-        let ValueConstructorVariant::Record {
-            name,
-            module,
-            variant_index,
-            arity,
-            ..
-        } = &constructor.variant
-        else {
-            return Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::ExpressionShape {
-                    kind: crate::planner::error::InvalidExpressionShapeKind::RecordConstructor,
-                },
-            });
-        };
-
-        if module != PRELUDE_MODULE_NAME {
-            let _linked_module = self.resolve_module_reference(module, name)?;
-        }
-        let constructor = self.custom_constructor_from_type(
-            constructor.type_.as_ref(),
-            name.clone(),
-            module,
-            usize::from(*variant_index),
-        )?;
-        self.validate_constructor_arity(constructor, usize::from(*arity))
-    }
-
-    pub(super) fn module_custom_constructor(
-        &self,
-        type_: &Type,
-        name: EcoString,
-        module: &EcoString,
-        variant_index: usize,
-        arity: usize,
-    ) -> Result<CustomConstructor, PlanError> {
-        let constructor = self.custom_constructor_from_type(type_, name, module, variant_index)?;
-        self.validate_constructor_arity(constructor, arity)
-    }
-
-    fn custom_constructor_from_type(
-        &self,
-        constructor_type: &Type,
-        name: EcoString,
-        module: &EcoString,
-        variant_index: usize,
-    ) -> Result<ResolvedCustomConstructor, PlanError> {
-        let signature = constructor_type.fn_types();
-        let return_type = match &signature {
-            Some((_, return_type)) => return_type.as_ref(),
-            None => constructor_type,
-        };
-        let mut type_parameters = self.type_parameters.clone();
-        let type_ = match ValueShape::from_gleam_in_with_external(
-            return_type,
-            &mut type_parameters,
-            &|name| self.registry.is_external_type(name),
-        ) {
-            ValueShape::Custom(shape) => shape.type_().clone(),
-            _ => {
-                return Err(PlanError::InvalidTypedAst {
-                    reason: InvalidTypedAstReason::CustomType {
-                        name,
-                        reason: crate::planner::error::InvalidCustomTypeReason::ConstructorType,
-                    },
-                });
-            }
-        };
-        let field_types = match signature {
-            Some((field_types, _)) => field_types,
-            None => Vec::new(),
-        };
-        let field_types = field_types
-            .into_iter()
-            .map(|field| {
-                ValueShape::from_gleam_in_with_external(
-                    field.as_ref(),
-                    &mut type_parameters,
-                    &|name| self.registry.is_external_type(name),
-                )
-                .value_type()
-            })
-            .collect();
-
-        self.custom_constructor_from_parts(type_, name, module, variant_index, field_types)
-    }
-
-    pub(super) fn custom_construction_shape(
-        &self,
-        construction: &crate::plan::CustomConstruction,
-    ) -> Result<CustomValueShape, PlanError> {
-        let constructor = construction.constructor();
-        let fields = construction.fields();
-        let type_ = constructor.type_();
-        let templates = if type_.type_name().package().is_empty()
-            && type_.type_name().module() == PRELUDE_MODULE_NAME
-            && type_.type_name().name() == "Result"
-        {
-            match constructor.index() {
-                0 => vec![CustomTypeTemplate::Parameter(
-                    crate::plan::CustomTypeParameterId(0),
-                )],
-                1 => vec![CustomTypeTemplate::Parameter(
-                    crate::plan::CustomTypeParameterId(1),
-                )],
-                _ => {
-                    return Err(invalid_custom_type(
-                        type_,
-                        InvalidCustomTypeReason::ConstructorIndex,
-                    ));
-                }
-            }
-        } else {
-            let definition = self
-                .custom_type_definition(type_.type_name())
-                .ok_or_else(|| {
-                    invalid_custom_type(type_, InvalidCustomTypeReason::UnknownDefinition)
-                })?;
-            definition
-                .constructor(constructor.index())
-                .ok_or_else(|| {
-                    invalid_custom_type(type_, InvalidCustomTypeReason::ConstructorIndex)
-                })?
-                .fields()
-                .iter()
-                .map(|field| field.type_().clone())
-                .collect()
-        };
-        let mut arguments = vec![None; type_.arguments().len()];
-        for (template, field) in templates.iter().zip(fields) {
-            if instantiate_custom_type_template(template, type_)? != field.value_type() {
-                return Err(invalid_custom_type(
-                    type_,
-                    InvalidCustomTypeReason::FieldType,
-                ));
-            }
-            collect_parameter_shapes(template, field.value_shape(), &mut arguments, type_)?;
-        }
-        let arguments = arguments
-            .into_iter()
-            .zip(type_.arguments())
-            .map(|(shape, type_)| match shape {
-                Some(shape) => shape,
-                None => ValueShape::from_value_type(type_.clone()),
-            })
-            .collect();
-        Ok(CustomValueShape::new(
-            type_.type_name().clone(),
-            arguments,
-            CustomConstructorRefinement::Exact(constructor.index()),
-        ))
-    }
-
-    pub(super) fn custom_expr_from_construction(
-        &self,
-        construction: crate::plan::CustomConstruction,
-    ) -> Result<crate::plan::CustomExpr, PlanError> {
-        let shape = self.custom_construction_shape(&construction)?;
-        Ok(crate::plan::CustomExpr::from_construction(
-            shape,
-            construction,
-        ))
-    }
-
-    pub(super) fn custom_pattern_constructor(
-        &self,
-        type_: &Type,
-        constructor: &PatternConstructor,
-        field_types: Vec<ValueType>,
-    ) -> Result<ResolvedCustomConstructor, PlanError> {
-        let mut type_parameters = self.type_parameters.clone();
-        let source_shape =
-            match ValueShape::from_gleam_in_with_external(type_, &mut type_parameters, &|name| {
-                self.registry.is_external_type(name)
-            }) {
-                ValueShape::Custom(shape) => shape,
-                _ => {
-                    return Err(PlanError::InvalidTypedAst {
-                        reason: InvalidTypedAstReason::CustomType {
-                            name: constructor.name.clone(),
-                            reason: InvalidCustomTypeReason::ConstructorType,
-                        },
-                    });
-                }
-            };
-        let mut constructor = self.custom_constructor_from_parts(
-            source_shape.type_().clone(),
-            constructor.name.clone(),
-            &constructor.module,
-            usize::from(constructor.constructor_index),
-            field_types,
-        )?;
-        constructor.source_shape = source_shape;
-        Ok(constructor)
-    }
-
-    pub(super) fn custom_field_access(
-        &self,
-        source: CustomExpr,
-        index: usize,
-        label: Option<EcoString>,
-        expected: &ValueType,
-    ) -> Result<(CustomFieldAccess, ValueShape), PlanError> {
-        let custom_type = source.type_();
-        let custom_shape = source.shape();
-        let Some(type_definition) = self.custom_type_definition(custom_type.type_name()) else {
-            return Err(invalid_custom_type(
-                custom_type,
-                InvalidCustomTypeReason::UnknownDefinition,
-            ));
-        };
-        if type_definition.parameters().len() != custom_type.arguments().len() {
-            return Err(invalid_custom_type(
-                custom_type,
-                InvalidCustomTypeReason::TypeArgumentCount,
-            ));
-        }
-
-        let constructors = match custom_shape.constructor() {
-            crate::plan::CustomConstructorRefinement::Exact(index) => {
-                vec![type_definition.constructor(index).ok_or_else(|| {
-                    invalid_custom_type(custom_type, InvalidCustomTypeReason::ConstructorIndex)
-                })?]
-            }
-            crate::plan::CustomConstructorRefinement::Any => {
-                type_definition.constructors().iter().collect()
-            }
-        };
-        let mut result_shape: Option<ValueShape> = None;
-        for constructor in constructors {
-            let field = constructor.fields().get(index).ok_or_else(|| {
-                invalid_custom_type(custom_type, InvalidCustomTypeReason::FieldIndex)
-            })?;
-            if field.label() != label.as_ref() {
-                return Err(invalid_custom_type(
-                    custom_type,
-                    InvalidCustomTypeReason::FieldLabel,
-                ));
-            }
-            let field_shape = instantiate_custom_shape_template(field.type_(), custom_shape)?;
-            let actual = field_shape.value_type();
-            if &actual != expected {
-                return Err(invalid_custom_type(
-                    custom_type,
-                    InvalidCustomTypeReason::FieldType,
-                ));
-            }
-            result_shape = Some(match result_shape {
-                Some(previous) => previous.merge(&field_shape).ok_or_else(|| {
-                    invalid_custom_type(custom_type, InvalidCustomTypeReason::FieldType)
-                })?,
-                None => field_shape,
-            });
-            constructor.fields().iter().try_for_each(|field| {
-                instantiate_custom_type_template(field.type_(), custom_type).map(|_| ())
-            })?;
-        }
-
-        let Some(result_shape) = result_shape else {
-            return Err(invalid_custom_type(
-                custom_type,
-                InvalidCustomTypeReason::FieldIndex,
-            ));
-        };
-        Ok((CustomFieldAccess::new(source, index, label), result_shape))
-    }
-
-    fn custom_constructor_from_parts(
-        &self,
-        type_: CustomType,
-        name: EcoString,
-        module: &EcoString,
-        variant_index: usize,
-        field_types: Vec<ValueType>,
-    ) -> Result<ResolvedCustomConstructor, PlanError> {
-        if module != type_.type_name().module() {
-            return Err(invalid_custom_type(
-                &type_,
-                InvalidCustomTypeReason::ConstructorModule,
-            ));
-        }
-        let (fields, constructor_count) = if type_.type_name().package().is_empty()
-            && module == PRELUDE_MODULE_NAME
-            && type_.type_name().module() == PRELUDE_MODULE_NAME
-            && type_.type_name().name() == "Result"
-        {
-            let [ok, error] = type_.arguments() else {
-                return Err(invalid_custom_type(
-                    &type_,
-                    InvalidCustomTypeReason::TypeArgumentCount,
-                ));
-            };
-            match variant_index {
-                0 if name == "Ok" => (vec![(None, ok.clone())], 2),
-                1 if name == "Error" => (vec![(None, error.clone())], 2),
-                0 | 1 => {
-                    return Err(invalid_custom_type(
-                        &type_,
-                        InvalidCustomTypeReason::ConstructorName,
-                    ));
-                }
-                _ => {
-                    return Err(invalid_custom_type(
-                        &type_,
-                        InvalidCustomTypeReason::ConstructorIndex,
-                    ));
-                }
-            }
-        } else {
-            let Some(type_definition) = self.custom_type_definition(type_.type_name()) else {
-                return Err(invalid_custom_type(
-                    &type_,
-                    InvalidCustomTypeReason::UnknownDefinition,
-                ));
-            };
-            if type_definition.parameters().len() != type_.arguments().len() {
-                return Err(invalid_custom_type(
-                    &type_,
-                    InvalidCustomTypeReason::TypeArgumentCount,
-                ));
-            }
-            let Some(constructor_definition) = type_definition.constructor(variant_index) else {
-                return Err(invalid_custom_type(
-                    &type_,
-                    InvalidCustomTypeReason::ConstructorIndex,
-                ));
-            };
-            if constructor_definition.name() != &name {
-                return Err(invalid_custom_type(
-                    &type_,
-                    InvalidCustomTypeReason::ConstructorName,
-                ));
-            };
-            if constructor_definition.fields().len() != field_types.len() {
-                return Err(invalid_custom_type(
-                    &type_,
-                    InvalidCustomTypeReason::ConstructorArity,
-                ));
-            }
-            let mut fields = Vec::with_capacity(constructor_definition.fields().len());
-            for field in constructor_definition.fields() {
-                fields.push((
-                    field.label().cloned(),
-                    instantiate_custom_type_template(field.type_(), &type_)?,
-                ));
-            }
-            (fields, type_definition.constructors().len())
-        };
-        if fields.len() != field_types.len()
-            || fields.iter().map(|(_, type_)| type_).ne(field_types.iter())
-        {
-            return Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::CustomType {
-                    name: type_.type_name().name().clone(),
-                    reason: InvalidCustomTypeReason::FieldType,
-                },
-            });
-        }
-        let fields = fields
-            .into_iter()
-            .map(|(label, type_)| CustomConstructorField::new(label, type_))
-            .collect();
-
-        let source_shape = CustomValueShape::any(type_.clone());
-        Ok(ResolvedCustomConstructor {
-            constructor: CustomConstructor::new(type_, name, variant_index, fields),
-            constructor_count,
-            source_shape,
-        })
-    }
-
     pub(super) fn lookup_function_local(
         &self,
         name: &EcoString,
@@ -3271,37 +2596,24 @@ mod tests {
     use super::FunctionLocalBinding;
     use super::{
         AnonymousFunctions, FunctionInfo, ModuleFunctionTarget, PlanContext, PlannedCapture,
-        ResolvedCustomConstructor, collect_parameter_shapes, instantiate_custom_shape_template,
-        instantiate_custom_type_template,
     };
     use crate::plan::{
         BitArrayFunctionLocalId, BitArrayListLocalId, BitArrayLocalId, BoolFunctionLocalId,
-        BoolListLocalId, BoolLocalId, CaptureArg, CustomConstruction, CustomConstructor,
-        CustomConstructorDefinition, CustomConstructorField, CustomConstructorRefinement,
-        CustomFieldDefinition, CustomFunctionLocal, CustomFunctionLocalId, CustomFunctionType,
-        CustomType, CustomTypeDefinition, CustomTypeName, CustomTypeParameterId,
-        CustomTypePublicity, CustomTypeTemplate, CustomValueShape, Expr, ExternalFunctionLocal,
+        BoolListLocalId, BoolLocalId, CaptureArg, CustomFunctionLocal, CustomFunctionLocalId,
+        CustomFunctionType, CustomType, CustomTypeName, ExternalFunctionLocal,
         ExternalFunctionLocalId, ExternalFunctionType, ExternalListLocalId, ExternalLocalId,
         ExternalTypeName, ExternalValueShape, FloatFunctionLocalId, FloatListLocalId, FloatLocalId,
-        FunctionExpr, FunctionFunctionLocal, FunctionFunctionLocalId, FunctionFunctionType,
-        FunctionListLocalId, FunctionReference, FunctionShape, FunctionType, GenericFunctionLocal,
-        GenericFunctionLocalId, GenericFunctionType, GenericListLocalId, IntExpr,
-        IntFunctionLocalId, IntListLocalId, IntLocalId, ListExpr, ListListLocalId, ListLocal,
-        ListLocalExpr, LocalId, NilFunctionLocalId, NilListLocalId, NilLocalId, ParamLocal,
-        ParamSlot, StringExpr, StringFunctionLocalId, StringListLocalId, StringLocalId,
-        TupleFunctionLocalId, TupleListLocalId, TupleLocalId, TypeParameterId,
-        UtfCodepointListLocalId, ValueShape, ValueType,
+        FunctionFunctionLocal, FunctionFunctionLocalId, FunctionFunctionType, FunctionListLocalId,
+        FunctionShape, FunctionType, GenericFunctionLocal, GenericFunctionLocalId,
+        GenericFunctionType, GenericListLocalId, IntFunctionLocalId, IntListLocalId, IntLocalId,
+        ListExpr, ListListLocalId, ListLocal, ListLocalExpr, LocalId, NilFunctionLocalId,
+        NilListLocalId, NilLocalId, ParamLocal, ParamSlot, StringFunctionLocalId,
+        StringListLocalId, StringLocalId, TupleFunctionLocalId, TupleListLocalId, TupleLocalId,
+        TypeParameterId, UtfCodepointListLocalId, ValueShape, ValueType,
     };
-    use crate::planner::{
-        InvalidCustomTypeReason, InvalidModuleReferenceReason, InvalidTypedAstReason, PlanError,
-    };
+    use crate::planner::{InvalidModuleReferenceReason, InvalidTypedAstReason, PlanError};
     use ecow::EcoString;
-    use gleam_core::ast::Publicity;
-    use gleam_core::type_::{
-        self, Deprecation, PatternConstructor, ValueConstructor, ValueConstructorVariant,
-    };
     use std::collections::HashMap;
-    use std::sync::Arc;
 
     fn custom_type() -> CustomType {
         CustomType::new(
@@ -3488,1215 +2800,6 @@ mod tests {
 
     fn function_template_id(function: FunctionInfo) -> crate::plan::FunctionTemplateId {
         function.signature.id()
-    }
-
-    #[test]
-    fn recursive_custom_shape_templates_reject_incompatible_parameter_shapes() {
-        let owner = CustomType::new(
-            CustomTypeName::new("geam".into(), "main".into(), "Generic".into()),
-            vec![ValueType::Int],
-        );
-        let owner_shape = CustomValueShape::new(
-            owner.type_name().clone(),
-            vec![ValueShape::Int],
-            CustomConstructorRefinement::Any,
-        );
-        let expected = PlanError::InvalidTypedAst {
-            reason: InvalidTypedAstReason::CustomType {
-                name: "Generic".into(),
-                reason: InvalidCustomTypeReason::ParameterType,
-            },
-        };
-
-        assert_eq!(
-            instantiate_custom_shape_template(
-                &CustomTypeTemplate::Parameter(CustomTypeParameterId(1)),
-                &owner_shape,
-            ),
-            Err(expected.clone()),
-        );
-        for template in [
-            CustomTypeTemplate::Tuple(vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(
-                1,
-            ))]),
-            CustomTypeTemplate::List(Box::new(CustomTypeTemplate::Parameter(
-                CustomTypeParameterId(1),
-            ))),
-            CustomTypeTemplate::Function {
-                arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(1))],
-                return_: Box::new(CustomTypeTemplate::Int),
-            },
-            CustomTypeTemplate::Function {
-                arguments: Vec::new(),
-                return_: Box::new(CustomTypeTemplate::Parameter(CustomTypeParameterId(1))),
-            },
-            CustomTypeTemplate::Custom {
-                name: CustomTypeName::new("geam".into(), "main".into(), "Nested".into()),
-                arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(1))],
-            },
-            CustomTypeTemplate::External {
-                name: ExternalTypeName::new("geam".into(), "main".into(), "Resource".into()),
-                arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(1))],
-            },
-        ] {
-            assert_eq!(
-                instantiate_custom_shape_template(&template, &owner_shape),
-                Err(expected.clone()),
-            );
-        }
-
-        let mut arguments = vec![None];
-        assert_eq!(
-            collect_parameter_shapes(
-                &CustomTypeTemplate::Parameter(CustomTypeParameterId(1)),
-                &ValueShape::Int,
-                &mut arguments,
-                &owner,
-            ),
-            Err(expected.clone()),
-        );
-        let mut arguments = vec![Some(ValueShape::Int)];
-        assert_eq!(
-            collect_parameter_shapes(
-                &CustomTypeTemplate::Parameter(CustomTypeParameterId(0)),
-                &ValueShape::String,
-                &mut arguments,
-                &owner,
-            ),
-            Err(expected.clone()),
-        );
-
-        let tuple = CustomTypeTemplate::Tuple(vec![CustomTypeTemplate::Parameter(
-            CustomTypeParameterId(0),
-        )]);
-        let mut arguments = vec![None];
-        assert_eq!(
-            collect_parameter_shapes(&tuple, &ValueShape::Int, &mut arguments, &owner),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &tuple,
-                &ValueShape::Tuple(vec![ValueShape::Int, ValueShape::String].into_boxed_slice(),),
-                &mut arguments,
-                &owner,
-            ),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &tuple,
-                &ValueShape::Tuple(vec![ValueShape::Int].into_boxed_slice()),
-                &mut arguments,
-                &owner,
-            ),
-            Ok(()),
-        );
-        assert_eq!(arguments, vec![Some(ValueShape::Int)]);
-
-        let module = EcoString::from("main");
-        let functions = HashMap::<EcoString, FunctionInfo>::new();
-        let mut anonymous = AnonymousFunctions::default();
-        let mut context = PlanContext::new(&module, &functions, &mut anonymous);
-        assert_eq!(
-            context.define_existing_param(
-                "value".into(),
-                &ParamLocal::int(IntLocalId(0)),
-                ValueShape::String,
-            ),
-            Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::ExpressionShape {
-                    kind: crate::planner::InvalidExpressionShapeKind::Invalid,
-                },
-            }),
-        );
-
-        for (template, shape) in [
-            (
-                CustomTypeTemplate::Tuple(vec![CustomTypeTemplate::Parameter(
-                    CustomTypeParameterId(1),
-                )]),
-                ValueShape::Tuple(vec![ValueShape::Int].into_boxed_slice()),
-            ),
-            (
-                CustomTypeTemplate::List(Box::new(CustomTypeTemplate::Parameter(
-                    CustomTypeParameterId(1),
-                ))),
-                ValueShape::List(Box::new(ValueShape::Int)),
-            ),
-            (
-                CustomTypeTemplate::Function {
-                    arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(1))],
-                    return_: Box::new(CustomTypeTemplate::Int),
-                },
-                ValueShape::Function(Box::new(FunctionShape::new(
-                    vec![ValueShape::Int],
-                    ValueShape::Int,
-                ))),
-            ),
-            (
-                CustomTypeTemplate::Function {
-                    arguments: Vec::new(),
-                    return_: Box::new(CustomTypeTemplate::Parameter(CustomTypeParameterId(1))),
-                },
-                ValueShape::Function(Box::new(FunctionShape::new(Vec::new(), ValueShape::Int))),
-            ),
-            (
-                CustomTypeTemplate::Custom {
-                    name: CustomTypeName::new("geam".into(), "main".into(), "Nested".into()),
-                    arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(1))],
-                },
-                ValueShape::Custom(CustomValueShape::new(
-                    CustomTypeName::new("geam".into(), "main".into(), "Nested".into()),
-                    vec![ValueShape::Int],
-                    CustomConstructorRefinement::Any,
-                )),
-            ),
-            (
-                CustomTypeTemplate::External {
-                    name: ExternalTypeName::new("geam".into(), "main".into(), "Resource".into()),
-                    arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(1))],
-                },
-                ValueShape::External(ExternalValueShape::new(
-                    ExternalTypeName::new("geam".into(), "main".into(), "Resource".into()),
-                    vec![ValueShape::Int],
-                )),
-            ),
-        ] {
-            let mut arguments = vec![None];
-            assert_eq!(
-                collect_parameter_shapes(&template, &shape, &mut arguments, &owner),
-                Err(expected.clone()),
-            );
-        }
-
-        let list = CustomTypeTemplate::List(Box::new(CustomTypeTemplate::Parameter(
-            CustomTypeParameterId(0),
-        )));
-        let mut arguments = vec![None];
-        assert_eq!(
-            collect_parameter_shapes(&list, &ValueShape::Int, &mut arguments, &owner),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &list,
-                &ValueShape::List(Box::new(ValueShape::Int)),
-                &mut arguments,
-                &owner,
-            ),
-            Ok(()),
-        );
-        assert_eq!(arguments, vec![Some(ValueShape::Int)]);
-
-        let external_name = ExternalTypeName::new("geam".into(), "main".into(), "Resource".into());
-        let external = CustomTypeTemplate::External {
-            name: external_name.clone(),
-            arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(0))],
-        };
-        let mut arguments = vec![None];
-        assert_eq!(
-            collect_parameter_shapes(&external, &ValueShape::Int, &mut arguments, &owner),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &external,
-                &ValueShape::External(ExternalValueShape::new(
-                    ExternalTypeName::new("geam".into(), "main".into(), "Other".into()),
-                    vec![ValueShape::Int],
-                )),
-                &mut arguments,
-                &owner,
-            ),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &external,
-                &ValueShape::External(ExternalValueShape::new(
-                    external_name.clone(),
-                    vec![ValueShape::Int, ValueShape::String],
-                )),
-                &mut arguments,
-                &owner,
-            ),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &external,
-                &ValueShape::External(ExternalValueShape::new(
-                    external_name,
-                    vec![ValueShape::Int],
-                )),
-                &mut arguments,
-                &owner,
-            ),
-            Ok(()),
-        );
-        assert_eq!(arguments, vec![Some(ValueShape::Int)]);
-
-        let function = CustomTypeTemplate::Function {
-            arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(0))],
-            return_: Box::new(CustomTypeTemplate::Parameter(CustomTypeParameterId(0))),
-        };
-        let mut arguments = vec![None];
-        assert_eq!(
-            collect_parameter_shapes(&function, &ValueShape::Int, &mut arguments, &owner),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &function,
-                &ValueShape::Function(Box::new(FunctionShape::new(Vec::new(), ValueShape::Int,))),
-                &mut arguments,
-                &owner,
-            ),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &function,
-                &ValueShape::Function(Box::new(FunctionShape::new(
-                    vec![ValueShape::Int],
-                    ValueShape::Int,
-                ))),
-                &mut arguments,
-                &owner,
-            ),
-            Ok(()),
-        );
-        assert_eq!(arguments, vec![Some(ValueShape::Int)]);
-
-        let nested_name = CustomTypeName::new("geam".into(), "main".into(), "Nested".into());
-        let custom = CustomTypeTemplate::Custom {
-            name: nested_name.clone(),
-            arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(0))],
-        };
-        let mut arguments = vec![None];
-        assert_eq!(
-            collect_parameter_shapes(&custom, &ValueShape::Int, &mut arguments, &owner),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &custom,
-                &ValueShape::Custom(CustomValueShape::new(
-                    CustomTypeName::new("geam".into(), "main".into(), "Other".into()),
-                    vec![ValueShape::Int],
-                    CustomConstructorRefinement::Any,
-                )),
-                &mut arguments,
-                &owner,
-            ),
-            Err(expected.clone()),
-        );
-        assert_eq!(
-            collect_parameter_shapes(
-                &custom,
-                &ValueShape::Custom(CustomValueShape::new(
-                    nested_name,
-                    vec![ValueShape::Int],
-                    CustomConstructorRefinement::Exact(0),
-                )),
-                &mut arguments,
-                &owner,
-            ),
-            Ok(()),
-        );
-        assert_eq!(arguments, vec![Some(ValueShape::Int)]);
-    }
-
-    #[test]
-    fn custom_shape_templates_preserve_every_primitive_shape() {
-        let owner_shape = CustomValueShape::new(
-            CustomTypeName::new("geam".into(), "main".into(), "Owner".into()),
-            vec![ValueShape::Int],
-            CustomConstructorRefinement::Any,
-        );
-
-        for (template, expected) in [
-            (CustomTypeTemplate::Int, ValueShape::Int),
-            (CustomTypeTemplate::Float, ValueShape::Float),
-            (CustomTypeTemplate::String, ValueShape::String),
-            (CustomTypeTemplate::BitArray, ValueShape::BitArray),
-            (CustomTypeTemplate::UtfCodepoint, ValueShape::UtfCodepoint),
-            (CustomTypeTemplate::Bool, ValueShape::Bool),
-            (CustomTypeTemplate::Nil, ValueShape::Nil),
-        ] {
-            assert_eq!(
-                instantiate_custom_shape_template(&template, &owner_shape),
-                Ok(expected),
-            );
-        }
-
-        let external_name = ExternalTypeName::new("geam".into(), "main".into(), "Resource".into());
-        assert_eq!(
-            instantiate_custom_shape_template(
-                &CustomTypeTemplate::External {
-                    name: external_name.clone(),
-                    arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(0))],
-                },
-                &owner_shape,
-            ),
-            Ok(ValueShape::External(ExternalValueShape::new(
-                external_name,
-                vec![ValueShape::Int],
-            ))),
-        );
-    }
-
-    #[test]
-    fn custom_construction_shape_rejects_malformed_constructor_metadata() {
-        let module = EcoString::from("main");
-        let functions = HashMap::<EcoString, FunctionInfo>::new();
-        let boxed_name = CustomTypeName::new("geam".into(), module.clone(), "Boxed".into());
-        let broken_name = CustomTypeName::new("geam".into(), module.clone(), "Broken".into());
-        let repeated_name = CustomTypeName::new("geam".into(), module.clone(), "Repeated".into());
-        let definitions = vec![
-            CustomTypeDefinition::new(
-                boxed_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                Vec::new(),
-                vec![CustomConstructorDefinition::new(
-                    "Boxed".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(None, CustomTypeTemplate::Int)],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                broken_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                vec![CustomTypeParameterId(0)],
-                vec![CustomConstructorDefinition::new(
-                    "Broken".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::Parameter(CustomTypeParameterId(1)),
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                repeated_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                vec![CustomTypeParameterId(0)],
-                vec![CustomConstructorDefinition::new(
-                    "Repeated".into(),
-                    0,
-                    vec![
-                        CustomFieldDefinition::new(
-                            None,
-                            CustomTypeTemplate::Parameter(CustomTypeParameterId(0)),
-                        ),
-                        CustomFieldDefinition::new(
-                            None,
-                            CustomTypeTemplate::Parameter(CustomTypeParameterId(0)),
-                        ),
-                    ],
-                )],
-            ),
-        ];
-        let mut anonymous = AnonymousFunctions::default();
-        let context =
-            PlanContext::new_with_custom_types(&module, &functions, &definitions, &mut anonymous);
-
-        let result = CustomType::new(
-            CustomTypeName::new(
-                "".into(),
-                type_::PRELUDE_MODULE_NAME.into(),
-                "Result".into(),
-            ),
-            vec![ValueType::Int, ValueType::String],
-        );
-        let invalid_result = CustomConstruction::try_new(
-            CustomConstructor::new(
-                result.clone(),
-                "Invalid".into(),
-                2,
-                vec![CustomConstructorField::new(None, ValueType::Int)],
-            ),
-            vec![Expr::int(IntExpr::value(1.into()))],
-        )
-        .expect("test construction has exact descriptor arity");
-        assert_eq!(
-            context.custom_construction_shape(&invalid_result),
-            Err(invalid_custom_constructor_error(
-                &result,
-                InvalidCustomTypeReason::ConstructorIndex,
-            )),
-        );
-        assert_eq!(
-            context
-                .custom_expr_from_construction(invalid_result)
-                .map(|_| ()),
-            Err(invalid_custom_constructor_error(
-                &result,
-                InvalidCustomTypeReason::ConstructorIndex,
-            )),
-        );
-
-        let error = CustomConstruction::try_new(
-            CustomConstructor::new(
-                result.clone(),
-                "Error".into(),
-                1,
-                vec![CustomConstructorField::new(None, ValueType::String)],
-            ),
-            vec![Expr::string(StringExpr::value("error".into()))],
-        )
-        .expect("test construction has exact descriptor arity");
-        assert_eq!(
-            context.custom_construction_shape(&error),
-            Ok(CustomValueShape::new(
-                result.type_name().clone(),
-                vec![ValueShape::Int, ValueShape::String],
-                CustomConstructorRefinement::Exact(1),
-            )),
-        );
-
-        let missing = CustomType::new(
-            CustomTypeName::new("geam".into(), module.clone(), "Missing".into()),
-            Vec::new(),
-        );
-        let unknown = CustomConstruction::try_new(
-            CustomConstructor::new(missing.clone(), "Missing".into(), 0, Vec::new()),
-            Vec::new(),
-        )
-        .expect("test construction has exact descriptor arity");
-        assert_eq!(
-            context.custom_construction_shape(&unknown),
-            Err(invalid_custom_constructor_error(
-                &missing,
-                InvalidCustomTypeReason::UnknownDefinition,
-            )),
-        );
-
-        let boxed = CustomType::new(boxed_name, Vec::new());
-        let invalid_constructor = CustomConstruction::try_new(
-            CustomConstructor::new(
-                boxed.clone(),
-                "Invalid".into(),
-                1,
-                vec![CustomConstructorField::new(None, ValueType::Int)],
-            ),
-            vec![Expr::int(IntExpr::value(1.into()))],
-        )
-        .expect("test construction has exact descriptor arity");
-        assert_eq!(
-            context.custom_construction_shape(&invalid_constructor),
-            Err(invalid_custom_constructor_error(
-                &boxed,
-                InvalidCustomTypeReason::ConstructorIndex,
-            )),
-        );
-
-        let invalid_field = CustomConstruction::try_new(
-            CustomConstructor::new(
-                boxed.clone(),
-                "Boxed".into(),
-                0,
-                vec![CustomConstructorField::new(None, ValueType::String)],
-            ),
-            vec![Expr::string(StringExpr::value("wrong".into()))],
-        )
-        .expect("test construction has exact descriptor arity");
-        assert_eq!(
-            context.custom_construction_shape(&invalid_field),
-            Err(invalid_custom_field_error(&boxed)),
-        );
-
-        let broken = CustomType::new(broken_name, vec![ValueType::Int]);
-        let invalid_template = CustomConstruction::try_new(
-            CustomConstructor::new(
-                broken.clone(),
-                "Broken".into(),
-                0,
-                vec![CustomConstructorField::new(None, ValueType::Int)],
-            ),
-            vec![Expr::int(IntExpr::value(1.into()))],
-        )
-        .expect("test construction has exact descriptor arity");
-        assert_eq!(
-            context.custom_construction_shape(&invalid_template),
-            Err(invalid_custom_constructor_error(
-                &broken,
-                InvalidCustomTypeReason::ParameterType,
-            )),
-        );
-
-        let choice = CustomType::new(
-            CustomTypeName::new("geam".into(), module.clone(), "Choice".into()),
-            Vec::new(),
-        );
-        let function_type =
-            FunctionType::new(vec![ValueType::Custom(choice.clone())], ValueType::Int);
-        let repeated = CustomType::new(
-            repeated_name,
-            vec![ValueType::Function(Box::new(function_type.clone()))],
-        );
-        let field = |local, constructor| {
-            let parameter_shape = CustomValueShape::new(
-                choice.type_name().clone(),
-                Vec::new(),
-                CustomConstructorRefinement::Exact(constructor),
-            );
-            let shape = FunctionShape::new(
-                vec![ValueShape::Custom(parameter_shape.clone())],
-                ValueShape::Int,
-            );
-            Expr::function(
-                FunctionExpr::reference(FunctionReference::new(
-                    crate::plan::monomorphic_function_instantiation(local, shape.clone()),
-                ))
-                .with_resolved_shape(shape)
-                .expect("function shape has the same nominal type"),
-            )
-        };
-        let conflicting_fields = CustomConstruction::try_new(
-            CustomConstructor::new(
-                repeated.clone(),
-                "Repeated".into(),
-                0,
-                vec![
-                    CustomConstructorField::new(
-                        None,
-                        ValueType::Function(Box::new(function_type.clone())),
-                    ),
-                    CustomConstructorField::new(None, ValueType::Function(Box::new(function_type))),
-                ],
-            ),
-            vec![field(0, 0), field(1, 1)],
-        )
-        .expect("test construction has exact descriptor arity");
-        assert_eq!(
-            context.custom_construction_shape(&conflicting_fields),
-            Err(invalid_custom_constructor_error(
-                &repeated,
-                InvalidCustomTypeReason::ParameterType,
-            )),
-        );
-    }
-
-    #[test]
-    fn custom_constructor_and_equality_metadata_reject_invalid_typed_ast_shapes() {
-        let module = EcoString::from("main");
-        let functions = HashMap::<EcoString, FunctionInfo>::new();
-        let generic_name = CustomTypeName::new("geam".into(), module.clone(), "Generic".into());
-        let function_name =
-            CustomTypeName::new("geam".into(), module.clone(), "FunctionBox".into());
-        let tuple_function_name =
-            CustomTypeName::new("geam".into(), module.clone(), "TupleFunctionBox".into());
-        let list_function_name =
-            CustomTypeName::new("geam".into(), module.clone(), "ListFunctionBox".into());
-        let broken_name = CustomTypeName::new("geam".into(), module.clone(), "Broken".into());
-        let tuple_broken_name =
-            CustomTypeName::new("geam".into(), module.clone(), "TupleBroken".into());
-        let custom_argument_broken_name =
-            CustomTypeName::new("geam".into(), module.clone(), "CustomArgumentBroken".into());
-        let nested_broken_name =
-            CustomTypeName::new("geam".into(), module.clone(), "NestedBroken".into());
-        let missing_name = CustomTypeName::new("geam".into(), module.clone(), "Missing".into());
-        let recursive_name = CustomTypeName::new("geam".into(), module.clone(), "Recursive".into());
-        let definitions = vec![
-            CustomTypeDefinition::new(
-                generic_name.clone(),
-                CustomTypePublicity::Public,
-                false,
-                vec![CustomTypeParameterId(0)],
-                vec![CustomConstructorDefinition::new(
-                    "Generic".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        Some("value".into()),
-                        CustomTypeTemplate::Parameter(CustomTypeParameterId(0)),
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                function_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                Vec::new(),
-                vec![CustomConstructorDefinition::new(
-                    "FunctionBox".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::Function {
-                            arguments: vec![CustomTypeTemplate::Int],
-                            return_: Box::new(CustomTypeTemplate::String),
-                        },
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                tuple_function_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                Vec::new(),
-                vec![CustomConstructorDefinition::new(
-                    "TupleFunctionBox".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::Tuple(vec![
-                            CustomTypeTemplate::Int,
-                            CustomTypeTemplate::Function {
-                                arguments: Vec::new(),
-                                return_: Box::new(CustomTypeTemplate::Nil),
-                            },
-                        ]),
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                list_function_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                Vec::new(),
-                vec![CustomConstructorDefinition::new(
-                    "ListFunctionBox".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::List(Box::new(CustomTypeTemplate::Function {
-                            arguments: Vec::new(),
-                            return_: Box::new(CustomTypeTemplate::Nil),
-                        })),
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                broken_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                vec![CustomTypeParameterId(0)],
-                vec![CustomConstructorDefinition::new(
-                    "Broken".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::Parameter(CustomTypeParameterId(1)),
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                tuple_broken_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                Vec::new(),
-                vec![CustomConstructorDefinition::new(
-                    "TupleBroken".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::Tuple(vec![CustomTypeTemplate::Parameter(
-                            CustomTypeParameterId(0),
-                        )]),
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                custom_argument_broken_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                Vec::new(),
-                vec![CustomConstructorDefinition::new(
-                    "CustomArgumentBroken".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::Custom {
-                            name: generic_name.clone(),
-                            arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(
-                                0,
-                            ))],
-                        },
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                nested_broken_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                Vec::new(),
-                vec![CustomConstructorDefinition::new(
-                    "NestedBroken".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::Custom {
-                            name: missing_name.clone(),
-                            arguments: Vec::new(),
-                        },
-                    )],
-                )],
-            ),
-            CustomTypeDefinition::new(
-                recursive_name.clone(),
-                CustomTypePublicity::Private,
-                false,
-                Vec::new(),
-                vec![CustomConstructorDefinition::new(
-                    "Recursive".into(),
-                    0,
-                    vec![CustomFieldDefinition::new(
-                        None,
-                        CustomTypeTemplate::Custom {
-                            name: recursive_name.clone(),
-                            arguments: Vec::new(),
-                        },
-                    )],
-                )],
-            ),
-        ];
-        let mut anonymous = AnonymousFunctions::default();
-        let context =
-            PlanContext::new_with_custom_types(&module, &functions, &definitions, &mut anonymous);
-        let generic_int = CustomType::new(generic_name.clone(), vec![ValueType::Int]);
-
-        for (template, expected) in [
-            (CustomTypeTemplate::Float, ValueType::Float),
-            (CustomTypeTemplate::BitArray, ValueType::BitArray),
-            (CustomTypeTemplate::UtfCodepoint, ValueType::UtfCodepoint),
-            (CustomTypeTemplate::Bool, ValueType::Bool),
-            (CustomTypeTemplate::Nil, ValueType::Nil),
-            (
-                CustomTypeTemplate::List(Box::new(CustomTypeTemplate::Int)),
-                ValueType::List(Box::new(ValueType::Int)),
-            ),
-        ] {
-            assert_eq!(
-                instantiate_custom_type_template(&template, &generic_int),
-                Ok(expected),
-            );
-        }
-
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                generic_int.clone(),
-                "Generic".into(),
-                &module,
-                0,
-                vec![ValueType::Int],
-            ),
-            Ok(ResolvedCustomConstructor {
-                constructor: CustomConstructor::new(
-                    generic_int.clone(),
-                    "Generic".into(),
-                    0,
-                    vec![CustomConstructorField::new(
-                        Some("value".into()),
-                        ValueType::Int,
-                    )],
-                ),
-                constructor_count: 1,
-                source_shape: CustomValueShape::any(generic_int.clone()),
-            }),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                generic_int.clone(),
-                "Generic".into(),
-                &module,
-                0,
-                vec![ValueType::String],
-            ),
-            Err(invalid_custom_field_error(&generic_int)),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                generic_int.clone(),
-                "Wrong".into(),
-                &module,
-                0,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_constructor_error(
-                &generic_int,
-                InvalidCustomTypeReason::ConstructorName,
-            )),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                generic_int.clone(),
-                "Generic".into(),
-                &"other".into(),
-                0,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_constructor_error(
-                &generic_int,
-                InvalidCustomTypeReason::ConstructorModule,
-            )),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                generic_int.clone(),
-                "Generic".into(),
-                &module,
-                0,
-                Vec::new(),
-            ),
-            Err(invalid_custom_constructor_error(
-                &generic_int,
-                InvalidCustomTypeReason::ConstructorArity,
-            )),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                generic_int.clone(),
-                "Generic".into(),
-                &module,
-                1,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_constructor_error(
-                &generic_int,
-                InvalidCustomTypeReason::ConstructorIndex,
-            )),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                CustomType::new(
-                    CustomTypeName::new("geam".into(), module.clone(), "Missing".into()),
-                    Vec::new(),
-                ),
-                "Missing".into(),
-                &module,
-                0,
-                Vec::new(),
-            ),
-            Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::CustomType {
-                    name: "Missing".into(),
-                    reason: InvalidCustomTypeReason::UnknownDefinition,
-                },
-            }),
-        );
-        let generic_without_arguments = CustomType::new(generic_name.clone(), Vec::new());
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                generic_without_arguments.clone(),
-                "Generic".into(),
-                &module,
-                0,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_constructor_error(
-                &generic_without_arguments,
-                InvalidCustomTypeReason::TypeArgumentCount,
-            )),
-        );
-
-        let result_type = CustomType::new(
-            CustomTypeName::new(
-                "".into(),
-                type_::PRELUDE_MODULE_NAME.into(),
-                "Result".into(),
-            ),
-            vec![ValueType::Int, ValueType::String],
-        );
-        assert_eq!(
-            ValueType::from_gleam(type_::result(type_::int(), type_::string()).as_ref()),
-            Some(ValueType::Custom(result_type.clone())),
-        );
-        let prelude = result_type.type_name().module().clone();
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                result_type.clone(),
-                "Error".into(),
-                &prelude,
-                1,
-                vec![ValueType::String],
-            ),
-            Ok(ResolvedCustomConstructor {
-                constructor: CustomConstructor::new(
-                    result_type.clone(),
-                    "Error".into(),
-                    1,
-                    vec![CustomConstructorField::new(None, ValueType::String)],
-                ),
-                constructor_count: 2,
-                source_shape: CustomValueShape::any(result_type.clone()),
-            }),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                result_type.clone(),
-                "Error".into(),
-                &prelude,
-                1,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_field_error(&result_type)),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                result_type.clone(),
-                "Error".into(),
-                &prelude,
-                0,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_constructor_error(
-                &result_type,
-                InvalidCustomTypeReason::ConstructorName,
-            )),
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                result_type.clone(),
-                "Ok".into(),
-                &prelude,
-                2,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_constructor_error(
-                &result_type,
-                InvalidCustomTypeReason::ConstructorIndex,
-            )),
-        );
-        let malformed_result_type =
-            CustomType::new(result_type.type_name().clone(), vec![ValueType::Int]);
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                malformed_result_type.clone(),
-                "Ok".into(),
-                &prelude,
-                0,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_constructor_error(
-                &malformed_result_type,
-                InvalidCustomTypeReason::TypeArgumentCount,
-            )),
-        );
-        let non_prelude_result = CustomType::new(
-            CustomTypeName::new(
-                "other".into(),
-                type_::PRELUDE_MODULE_NAME.into(),
-                "Result".into(),
-            ),
-            vec![ValueType::Int, ValueType::String],
-        );
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                non_prelude_result.clone(),
-                "Ok".into(),
-                &prelude,
-                0,
-                vec![ValueType::Int],
-            ),
-            Err(invalid_custom_constructor_error(
-                &non_prelude_result,
-                InvalidCustomTypeReason::UnknownDefinition,
-            )),
-        );
-        let broken = CustomType::new(broken_name, vec![ValueType::Int]);
-        assert_eq!(
-            context.custom_constructor_from_parts(
-                broken.clone(),
-                "Broken".into(),
-                &module,
-                0,
-                vec![ValueType::Int],
-            ),
-            Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::CustomType {
-                    name: "Broken".into(),
-                    reason: InvalidCustomTypeReason::ParameterType,
-                },
-            }),
-        );
-
-        let missing_parameter = CustomType::new(generic_name, Vec::new());
-        let parameter_error = Err(PlanError::InvalidTypedAst {
-            reason: InvalidTypedAstReason::CustomType {
-                name: "Generic".into(),
-                reason: InvalidCustomTypeReason::ParameterType,
-            },
-        });
-        assert_eq!(
-            instantiate_custom_type_template(
-                &CustomTypeTemplate::Tuple(vec![CustomTypeTemplate::Parameter(
-                    CustomTypeParameterId(0),
-                )]),
-                &missing_parameter,
-            ),
-            parameter_error.clone(),
-        );
-        assert_eq!(
-            instantiate_custom_type_template(
-                &CustomTypeTemplate::List(Box::new(CustomTypeTemplate::Parameter(
-                    CustomTypeParameterId(0),
-                ))),
-                &missing_parameter,
-            ),
-            parameter_error.clone(),
-        );
-        assert_eq!(
-            instantiate_custom_type_template(
-                &CustomTypeTemplate::Function {
-                    arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(0))],
-                    return_: Box::new(CustomTypeTemplate::Int),
-                },
-                &missing_parameter,
-            ),
-            parameter_error.clone(),
-        );
-        assert_eq!(
-            instantiate_custom_type_template(
-                &CustomTypeTemplate::Function {
-                    arguments: Vec::new(),
-                    return_: Box::new(CustomTypeTemplate::Parameter(CustomTypeParameterId(0))),
-                },
-                &missing_parameter,
-            ),
-            parameter_error.clone(),
-        );
-        assert_eq!(
-            instantiate_custom_type_template(
-                &CustomTypeTemplate::Custom {
-                    name: missing_parameter.type_name().clone(),
-                    arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(0))],
-                },
-                &missing_parameter,
-            ),
-            parameter_error.clone(),
-        );
-        assert_eq!(
-            instantiate_custom_type_template(
-                &CustomTypeTemplate::External {
-                    name: crate::plan::ExternalTypeName::new(
-                        "dependency".into(),
-                        "dependency/resource".into(),
-                        "Resource".into(),
-                    ),
-                    arguments: vec![CustomTypeTemplate::Parameter(CustomTypeParameterId(0))],
-                },
-                &missing_parameter,
-            ),
-            parameter_error,
-        );
-    }
-
-    #[test]
-    fn custom_constructor_typed_ast_margins_are_exact() {
-        let module = EcoString::from("main");
-        let functions = HashMap::<EcoString, FunctionInfo>::new();
-        let mut anonymous = AnonymousFunctions::default();
-        let context = PlanContext::new(&module, &functions, &mut anonymous);
-        let local = ValueConstructor::local_variable(
-            crate::planner::support::dummy_span(),
-            gleam_core::type_::error::VariableOrigin::generated(),
-            type_::int(),
-        );
-        assert_eq!(
-            context.custom_constructor(&local),
-            Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::ExpressionShape {
-                    kind: crate::planner::InvalidExpressionShapeKind::RecordConstructor,
-                },
-            }),
-        );
-        let invalid_field = ValueConstructor {
-            publicity: Publicity::Private,
-            deprecation: Deprecation::NotDeprecated,
-            type_: Arc::new(gleam_core::type_::Type::Fn {
-                arguments: vec![type_::generic_var(0)],
-                return_: type_::result(type_::int(), type_::string()),
-            }),
-            variant: ValueConstructorVariant::Record {
-                name: "Ok".into(),
-                arity: 1,
-                field_map: None,
-                location: crate::planner::support::dummy_span(),
-                module: type_::PRELUDE_MODULE_NAME.into(),
-                variants_count: 2,
-                variant_index: 0,
-                documentation: None,
-            },
-        };
-        assert_eq!(
-            context.custom_constructor(&invalid_field),
-            Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::CustomType {
-                    name: "Result".into(),
-                    reason: InvalidCustomTypeReason::FieldType,
-                },
-            }),
-        );
-        let invalid_constructor_type = ValueConstructor {
-            type_: type_::int(),
-            ..invalid_field
-        };
-        assert_eq!(
-            context.custom_constructor(&invalid_constructor_type),
-            Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::CustomType {
-                    name: "Ok".into(),
-                    reason: InvalidCustomTypeReason::ConstructorType,
-                },
-            }),
-        );
-
-        let pattern = PatternConstructor {
-            name: "Invalid".into(),
-            field_map: None,
-            documentation: None,
-            module: module.clone(),
-            location: crate::planner::support::dummy_span(),
-            constructor_index: 0,
-        };
-        assert_eq!(
-            context.custom_pattern_constructor(
-                type_::generic_var(0).as_ref(),
-                &pattern,
-                Vec::new()
-            ),
-            Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::CustomType {
-                    name: "Invalid".into(),
-                    reason: InvalidCustomTypeReason::ConstructorType,
-                },
-            }),
-        );
-        assert_eq!(
-            context.custom_pattern_constructor(type_::int().as_ref(), &pattern, Vec::new()),
-            Err(PlanError::InvalidTypedAst {
-                reason: InvalidTypedAstReason::CustomType {
-                    name: "Invalid".into(),
-                    reason: InvalidCustomTypeReason::ConstructorType,
-                },
-            }),
-        );
-    }
-
-    fn invalid_custom_constructor_error(
-        type_: &CustomType,
-        reason: InvalidCustomTypeReason,
-    ) -> PlanError {
-        PlanError::InvalidTypedAst {
-            reason: InvalidTypedAstReason::CustomType {
-                name: type_.type_name().name().clone(),
-                reason,
-            },
-        }
-    }
-
-    fn invalid_custom_field_error(type_: &CustomType) -> PlanError {
-        PlanError::InvalidTypedAst {
-            reason: InvalidTypedAstReason::CustomType {
-                name: type_.type_name().name().clone(),
-                reason: InvalidCustomTypeReason::FieldType,
-            },
-        }
     }
 
     #[test]
