+++ src/planner/expression/var.rs
@@ -644,8 +644,10 @@ pub fn main() {
             plan_module(record_constructor),
             Err(PlanError::InvalidTypedAst {
                 reason: InvalidTypedAstReason::CustomType {
+                    package: "geam".into(),
+                    module: "main".into(),
                     name: "Boxed".into(),
-                    reason: InvalidCustomTypeReason::UnknownDefinition,
+                    reason: Box::new(InvalidCustomTypeReason::MissingDefinition),
                 },
             }),
         );
@@ -673,8 +675,13 @@ pub fn main() { Boxed }
             plan_module(constructor_arity_mismatch),
             Err(PlanError::InvalidTypedAst {
                 reason: InvalidTypedAstReason::CustomType {
+                    package: "geam".into(),
+                    module: "main".into(),
                     name: "Boxed".into(),
-                    reason: InvalidCustomTypeReason::ConstructorArity,
+                    reason: Box::new(InvalidCustomTypeReason::ConstructorArity {
+                        expected: 1,
+                        actual: 0,
+                    }),
                 },
             }),
         );
@@ -747,8 +754,13 @@ pub fn main() {
             plan_module(arity_mismatch),
             Err(PlanError::InvalidTypedAst {
                 reason: InvalidTypedAstReason::CustomType {
+                    package: "geam".into(),
+                    module: "main".into(),
                     name: "Boxed".into(),
-                    reason: InvalidCustomTypeReason::ConstructorArity,
+                    reason: Box::new(InvalidCustomTypeReason::ConstructorArity {
+                        expected: 1,
+                        actual: 0,
+                    }),
                 },
             }),
         );
@@ -762,8 +774,12 @@ pub fn main() {
             plan_module(constructor_type_mismatch),
             Err(PlanError::InvalidTypedAst {
                 reason: InvalidTypedAstReason::CustomType {
+                    package: "".into(),
+                    module: "main".into(),
                     name: "Boxed".into(),
-                    reason: InvalidCustomTypeReason::ConstructorType,
+                    reason: Box::new(InvalidCustomTypeReason::ConstructorType {
+                        actual: crate::plan::ValueType::Int,
+                    }),
                 },
             }),
         );
